refactor: change job payload validation to use joi
I didn't add a schema for every job yet. We can do that as we go.
I decided that to be consistent, a job payload must be either undefined or an object.
In the process of writing the schema's, I had to fix a few issues, like exeting subscribeMany with an empty array
Validation error messages still have to be improved